+++ mcp-types/src/MCP/Aeson.hs
@@ -1,6 +1,5 @@
 {- |
 Module:      MCP.Aeson
-Copyright:   (c) DPella AB 2025
 License:     MPL-2.0
 Maintainer:  <matti@dpella.io>, <lobo@dpella.io>
 -}